Move Scroller into Things folder
